add bleed sleep punisher test

+++ core/log.py
@@ -3,7 +3,7 @@ import sys
 
 loglevel = None
 if len(sys.argv) >= 2:
-    loglevel = int(sys.argv[1])
+    loglevel = float(sys.argv[1])
 
 #g_log = []
 g_log_active = []
